feat: min-w 값 1280px 기준으로 설정 및 폴더 구조 정리
