+++ client/src/components/mainpage/layout/DashboardLayout.tsx
@@ -1,15 +1,15 @@
 import { Dashboard } from '@/components/mainpage/dashboard/Dashboard';
 import { StatsDashboard } from '@/components/mainpage/dashboard/StatsDashboard';
-import { AutomationStatus } from '@/components/mainpage/dashboard/Automation';
-import { MenuContainer } from '../sidebar/MenuContainer';
+import { MenuContainer } from '../dashboard/MenuContainer';
+import { AutomationContainer } from '../dashboard/AutomationContainer';
 
 export const DashboardLayout = () => {
   return (
-    <main className="flex-1 p-5">
+    <main className="flex-1 p-5 min-w-[800px]">
       <div className="flex flex-col gap-8">
         <div className="flex gap-5">
           <MenuContainer />
-          <AutomationStatus />
+          <AutomationContainer />
         </div>
         <StatsDashboard />
         <Dashboard />
