+++ client/src/components/mainpage/layout/InfoSidebarLayout.tsx
@@ -4,7 +4,7 @@ import { OrganizationContainer } from "../sidebar/OrganizationContainer";
 
 export const InfoSidebarLayout = () => {
   return (
-    <div className={`sidebar-right-width min-h-[930px] bg-page-sidebar-bg p-1 shadow-xl border rounded-[10px] space-y-6`}>
+    <div className={`sidebar-right-width min-w-[293px] min-h-[930px] bg-page-sidebar-bg p-1 shadow-xl border rounded-[10px] space-y-6`}>
       <NotificationContainer />
       <ScheduleContainer />
       <OrganizationContainer />
